IDL defined exceptions now inherit from StandardError
This allows them to be used as normal Ruby exceptions.
