+++ lib/sparsam.rb
@@ -4,6 +4,7 @@ require 'sparsam_native'
 require 'sparsam/types'
 require 'sparsam/exceptions'
 require 'sparsam/struct'
+require 'sparsam/application_exception'
 require 'sparsam/union'
 require 'sparsam/deserializer'
